Revert "fix for omnipath: changing gaspi_notification_id from unsigned int to byte length, because omnipath writes single bytes at a time, which can result in races because (a) unexpected notification value is read out (b) zero value after reset is filled, bug fix: restrict the notification value to byte length which can not be overwritten: fix has been tested on Omnipath system at kraken.cerfacs.fr"
This reverts commit 7b65525c6823fad1559f168c2b04bc43159263b3.

+++ include/GASPI.h
@@ -29,8 +29,6 @@ along with GPI-2. If not, see <http://www.gnu.org/licenses/>.
 #ifndef GPI2_H
 #define GPI2_H
 
-#include <stdint.h>
-
 #ifdef __cplusplus
 extern "C"
 {
@@ -63,7 +61,7 @@ extern "C"
   typedef float gaspi_time_t;
   typedef unsigned long gaspi_cycles_t;
   typedef unsigned short gaspi_notification_id_t;
-  typedef uint8_t gaspi_notification_t;
+  typedef unsigned int gaspi_notification_t;
   typedef unsigned int gaspi_statistic_counter_t;
   typedef char * gaspi_string_t;
